gin/user/userservice_test.go
package user_test

import (
	"fmt"
	"testing"
)

// UserServiceTest is a test suite for the UserService.
type UserServiceTest struct {
	service *UserService
}

// SetupTest sets up the test environment.
func (suite *UserServiceTest) SetupTest() {
	// Initialize the database and other dependencies
	// ...

	// Create a new UserService instance
	suite.service = &UserService{}
}

// TestLogin tests the Login function.
func (suite *UserServiceTest) TestLogin(t *testing.T) {
	// Test case 1: user not exist
	t.Run("UserNotExist", func(t *testing.T) {
		token, err := suite.service.Login("nonexistentuser", "password123")

		// Verify the error message
		expectedErr := fmt.Errorf("user not exist")
		if err.Error() != expectedErr.Error() {
			t.Errorf("expected error: %v, got: %v", expectedErr, err)
		}

		// Verify the token is empty
		if token != "" {
			t.Errorf("expected empty token, got: %s", token)
		}
	})

	// Test case 2: password error
	t.Run("PasswordError", func(t *testing.T) {
		// Create a test user in the database with wrong password
		// ...

		token, err := suite.service.Login("testuser", "wrongpassword")

		// Verify the error message
		expectedErr := fmt.Errorf("password error")
		if err.Error() != expectedErr.Error() {
			t.Errorf("expected error: %v, got: %v", expectedErr, err)
		}

		// Verify the token is empty
		if token != "" {
			t.Errorf("expected empty token, got: %s", token)
		}
	})

	// Test case 3: successful login
	t.Run("Success", func(t *testing.T) {
		// Create a test user in the database with correct password
		// ...

		token, err := suite.service.Login("testuser", "correctpassword")

		// Verify there is no error
		if err != nil {
			t.Errorf("unexpected error: %v", err)
		}

		// Verify the token is generated
		if token == "" {
			t.Errorf("expected non-empty token, got: %s", token)
		}
	})
}

// TearDownTest tears down the test environment.
func (suite *UserServiceTest) TearDownTest() {
	// Close the database and other dependencies
	// ...
}

// RunUserServiceTest runs all the test cases in UserServiceTest.
func RunUserServiceTest(t *testing.T) {
	suite := &UserServiceTest{}
	suite.SetupTest()
	defer suite.TearDownTest()

	t.Run("UserService", suite.TestLogin)
}
